Refactors the getStarred function, refactors its tests.
getStarred queries the Subsonic server and caches the request. It handles both GET and JSONP. It returns meaningful messages when there is a problem and distinguishes between HTTP errors and Subsonic errors.
Those messages are meant to be displayed by the controller using the notifications.

The tests are refactored too : most of what is common is at the top describe to keep the tests lean. Tests for error conditions have been added.