Refactors subsonic's getSongs(), adds a recursiveGetSongs()
So that the separation between subsonic-service and subsonic-controller is clear: the service makes requests and deals with $http, the controller displays, add to the player queue, etc.
I've also added a recursive version of getSongs() which enables us to play all the songs in a directory recursively, e.g. all the songs of an artist regardless of their album.

- Adds unit-tests to the controller and service for getSongs() and recursiveGetSongs()